refactor: 하단 버튼 모듈화
